db: 하이라이트 인덱스 관련 데이터 타입 int로 수정

+++ backend/src/main/resources/db/migration/V3__create_highlight_table.sql
@@ -7,9 +7,9 @@ CREATE TABLE highlight
     questionId      BIGINT NOT NULL,
     review_group_id BIGINT NOT NULL,
     answer_id       BIGINT NOT NULL,
-    line_index      BIGINT NOT NULL,
-    start_index     BIGINT NOT NULL,
-    end_index       BIGINT NOT NULL,
+    line_index      INT NOT NULL,
+    start_index     INT NOT NULL,
+    end_index       INT NOT NULL,
     PRIMARY KEY (id)
 );
